+++ src/Types/bookingTestArray.ts
@@ -1,0 +1,550 @@
+import { BookingArray } from "./calendar";
+
+export const bookings: BookingArray = [
+  {
+    "2023-12-12": {
+      PC1: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC2: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC3: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC4: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC5: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+    },
+    "2023-12-13": {
+      PC1: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC2: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC3: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC4: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+      PC5: [
+        {
+          time: "14.00",
+          booked: false,
+        },
+        {
+          time: "14.30",
+          booked: false,
+        },
+        {
+          time: "15.00",
+          booked: false,
+        },
+        {
+          time: "15.30",
+          booked: false,
+        },
+        {
+          time: "16.00",
+          booked: false,
+        },
+        {
+          time: "16.30",
+          booked: false,
+        },
+        {
+          time: "17.00",
+          booked: false,
+        },
+        {
+          time: "17.30",
+          booked: false,
+        },
+        {
+          time: "18.00",
+          booked: false,
+        },
+        {
+          time: "18.30",
+          booked: false,
+        },
+        {
+          time: "19.00",
+          booked: false,
+        },
+        {
+          time: "19.30",
+          booked: false,
+        },
+        {
+          time: "20.00",
+          booked: false,
+        },
+      ],
+    },
+  },
+];
