+++ src/components/ui/calendar.tsx
@@ -7,61 +7,61 @@ import { DayPicker } from "react-day-picker";
 import { cn } from "../../lib/utils";
 import { buttonVariants } from "../../components/ui/button";
 
-const disabledDays2 = (numberOfDays: number): Date[] => {
-  const disabledDays: Date[] = [];
-  const daysInWeek = 7;
+// const disabledDays2 = (numberOfDays: number): Date[] => {
+//   const disabledDays: Date[] = [];
+//   const daysInWeek = 7;
 
-  for (let i = 0; i < numberOfDays; i++) {
-    const currentDate = new Date();
-    currentDate.setDate(currentDate.getDate() + i);
+//   for (let i = 0; i < numberOfDays; i++) {
+//     const currentDate = new Date();
+//     currentDate.setDate(currentDate.getDate() + i);
 
-    const dayOfWeek = currentDate.getDay();
-    // 5 corresponds to Friday and 6 corresponds to Saturday
-    if (dayOfWeek === 5 || dayOfWeek === 6) {
-      const year = currentDate.getFullYear();
-      const month = String(currentDate.getMonth() + 1).padStart(2, "0"); // Months are zero-indexed
-      const day = String(currentDate.getDate()).padStart(2, "0");
+//     const dayOfWeek = currentDate.getDay();
+//     // 5 corresponds to Friday and 6 corresponds to Saturday
+//     if (dayOfWeek === 5 || dayOfWeek === 6) {
+//       const year = currentDate.getFullYear();
+//       const month = String(currentDate.getMonth() + 1).padStart(2, "0"); // Months are zero-indexed
+//       const day = String(currentDate.getDate()).padStart(2, "0");
 
-      // Format the date as "YYYY, MM, DD"
-      const formattedDate = `${Number(year)}, ${Number(month)}, ${Number(day)}`;
+//       // Format the date as "YYYY, MM, DD"
+//       const formattedDate = `${Number(year)}, ${Number(month)}, ${Number(day)}`;
 
-      disabledDays.push(new Date(formattedDate));
-    }
-  }
-  const futureDate = new Date();
-  futureDate.setDate(futureDate.getDate() + 14);
-  const fYear = futureDate.getFullYear();
-  const fMonth = String(futureDate.getMonth() + 1).padStart(2, "0"); // Months are zero-indexed
-  const fDay = String(futureDate.getDate()).padStart(2, "0");
+//       disabledDays.push(new Date(formattedDate));
+//     }
+//   }
+//   const futureDate = new Date();
+//   futureDate.setDate(futureDate.getDate() + 14);
+//   const fYear = futureDate.getFullYear();
+//   const fMonth = String(futureDate.getMonth() + 1).padStart(2, "0"); // Months are zero-indexed
+//   const fDay = String(futureDate.getDate()).padStart(2, "0");
 
-  // Format the date as "YYYY, MM, DD"
-  const fFormattedDate = `${Number(fYear)}, ${Number(fMonth)}, ${Number(fDay)}`;
-  //@ts-ignore
-  disabledDays.push({ from: new Date(fFormattedDate), to: new Date(2050, 1, 1) });
+//   // Format the date as "YYYY, MM, DD"
+//   const fFormattedDate = `${Number(fYear)}, ${Number(fMonth)}, ${Number(fDay)}`;
+//   //@ts-ignore
+//   disabledDays.push({ from: new Date(fFormattedDate), to: new Date(2050, 1, 1) });
 
-  //past day
-  const pastDate = new Date();
-  pastDate.setDate(pastDate.getDate() - 1);
-  const pYear = pastDate.getFullYear();
-  const pMonth = String(pastDate.getMonth() + 1).padStart(2, "0"); // Months are zero-indexed
-  const pDay = String(pastDate.getDate()).padStart(2, "0");
+//   //past day
+//   const pastDate = new Date();
+//   pastDate.setDate(pastDate.getDate() - 1);
+//   const pYear = pastDate.getFullYear();
+//   const pMonth = String(pastDate.getMonth() + 1).padStart(2, "0"); // Months are zero-indexed
+//   const pDay = String(pastDate.getDate()).padStart(2, "0");
 
-  const pFormattedDate = `${Number(pYear)}, ${Number(pMonth)}, ${Number(pDay)}`;
-  //@ts-ignore
-  disabledDays.push({ from: new Date(2023, 1, 1), to: new Date(pFormattedDate) });
+//   const pFormattedDate = `${Number(pYear)}, ${Number(pMonth)}, ${Number(pDay)}`;
+//   //@ts-ignore
+//   disabledDays.push({ from: new Date(2023, 1, 1), to: new Date(pFormattedDate) });
 
-  console.log(disabledDays);
+//   console.log(disabledDays);
 
-  return disabledDays;
-};
+//   return disabledDays;
+// };
 
-const disabledDays = disabledDays2(14);
+// const disabledDays = disabledDays2(14);
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>;
-function Calendar({ className, classNames, showOutsideDays = true, ...props }: CalendarProps) {
+function Calendar({ className, classNames, showOutsideDays = true, disabled, ...props }: CalendarProps) {
   return (
     <DayPicker
-      disabled={disabledDays}
+      disabled={disabled}
       defaultMonth={new Date()}
       showOutsideDays={showOutsideDays}
       weekStartsOn={1}
